fix(en.qiscans): move to new template
