+++ sources/en.qiscans/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 
 [dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs.git", version = "0.3.0" }
-iken = { path = "../../templates/iken" }
+ezmanhwa = { path = "../../templates/ezmanhwa" }
 
 [dev-dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs.git", features = ["test"] }
