+++ sources/en.qiscans/res/filters.json
@@ -1,83 +1,141 @@
 [
 	{
-		"type": "sort",
-		"id": "orderBy",
-		"canAscend": false,
-		"options": ["Created at", "Updated at", "Views", "Title"]
+		"type": "select",
+		"id": "sort",
+		"title": "Sort By",
+		"options": ["Default", "Popular", "Latest"],
+		"ids": ["", "popular", "latest"]
 	},
 	{
 		"type": "select",
-		"id": "seriesStatus",
+		"id": "status",
 		"title": "Status",
-		"options": ["All", "Ongoing", "Completed", "Dropped", "Hiatus"],
+		"options": ["Any", "Ongoing", "Completed", "Dropped", "Hiatus"],
 		"ids": ["", "ONGOING", "COMPLETED", "DROPPED", "HIATUS"]
 	},
 	{
-		"type": "multi-select",
+		"type": "select",
+		"id": "type",
+		"title": "Type",
+		"options": ["Any", "Manhwa", "Manga", "Manhua", "Novel"],
+		"ids": ["", "MANHWA", "MANGA", "MANHUA", "NOVEL"]
+	},
+	{
+		"type": "select",
+		"id": "genre",
 		"title": "Genres",
 		"isGenre": true,
-		"canExclude": false,
 		"options": [
-			"Slice of Life",
-			"Tragedy",
-			"Supernatural",
-			"Romance",
-			"Josei",
-			"Mature",
-			"Cooking",
-			"Ecchi",
-			"Horror",
-			"hidden",
-			"Historical",
-			"Comedy",
-			"Reincarnation",
-			"Harem",
-			"Gender Bender",
-			"Seinen",
-			"School Life",
-			"Martial Arts",
-			"Sports",
-			"Psychological",
-			"Virtual Reality",
-			"Mystery",
-			"Sci-Fi",
+			"All",
 			"Adventure",
-			"cultivation",
+			"Adventure (Alt)",
+			"Live",
+			"Acting",
 			"Action",
-			"Shounen",
+			"Action (Alt)",
+			"Comedy",
+			"Cooking",
+			"Crazy MC",
+			"Drama",
+			"Ecchi",
 			"Fantasy",
-			"Drama"
+			"Fantasy (Alt)",
+			"Fight",
+			"Gender Bender",
+			"Harem",
+			"Historical",
+			"Horror",
+			"Josei",
+			"Manhua",
+			"Martial Arts",
+			"Mature",
+			"Mecha",
+			"Murim",
+			"Mystery",
+			"Psychological",
+			"Reincarnation",
+			"Revenge",
+			"Romance",
+			"School Life",
+			"Sci-Fi",
+			"Seinen",
+			"Shounen",
+			"Slice of Life",
+			"Sports",
+			"Supernatural",
+			"Tragedy",
+			"Virtual Reality",
+			"Apocalypse",
+			"Cultivation",
+			"Hidden",
+			"Magic",
+			"Medieval Area",
+			"Munchkin",
+			"Myth",
+			"Politics",
+			"Superpower",
+			"System",
+			"Taming",
+			"Tower",
+			"Urban",
+			"Vampires",
+			"Wuxia"
 		],
 		"ids": [
-			"9",
-			"16",
-			"18",
-			"19",
-			"20",
-			"28",
-			"34",
-			"37",
-			"38",
-			"106",
-			"1",
-			"8",
-			"11",
-			"145",
-			"146",
-			"6",
-			"17",
-			"5",
-			"149",
-			"14",
-			"191",
-			"55",
-			"24",
-			"7",
-			"105",
-			"2",
-			"15",
-			"3",
-			"4"
+			"",
+			"adventure",
+			"adventure-589",
+			"live",
+			"acting",
+			"action",
+			"action-582",
+			"comedy",
+			"cooking",
+			"crazy-mc",
+			"drama",
+			"ecchi",
+			"fantasy",
+			"fantasy-747",
+			"fight",
+			"gender-bender",
+			"harem",
+			"historical",
+			"horror",
+			"josei",
+			"manhua",
+			"martial-arts",
+			"mature",
+			"mecha",
+			"murim",
+			"mystery",
+			"psychological",
+			"reincarnation",
+			"revenge",
+			"romance",
+			"school-life",
+			"sci-fi",
+			"seinen",
+			"shounen",
+			"slice-of-life",
+			"sports",
+			"supernatural",
+			"tragedy",
+			"virtual-reality",
+			"apocalypce",
+			"cultivation",
+			"hidden",
+			"magic",
+			"medieval-area",
+			"munchkin",
+			"myth",
+			"politics",
+			"superpower",
+			"system",
+			"taming",
+			"tower",
+			"urban",
+			"vampiers",
+			"wuxia"
 		]
 	}
 ]
