+++ templates/ezmanhwa/Cargo.toml
@@ -1,24 +1,12 @@
 [package]
-name = "ezmanga"
+name = "ezmanhwa"
 version = "0.1.0"
 edition = "2024"
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs.git", package = "aidoku", features = ["json"], version = "0.3.0" }
-ezmanhwa = { path = "../../templates/ezmanhwa" }
+serde = { version = "1.0", default-features = false, features = ["derive", "alloc"] }
 
 [dev-dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs.git", features = ["test"] }
 aidoku-test = { git = "https://github.com/Aidoku/aidoku-rs.git" }
-
-[profile.dev]
-panic = "abort"
-
-[profile.release]
-panic = "abort"
-opt-level = "s"
-strip = true
-lto = true
